+++ lua/tree-sitter/grammar.js
@@ -11,8 +11,8 @@ const PREC = {
   VAR_COMMENT_PREFIX: 2,
   BODY_PREFIX: 2,
   RAW_BODY: 3,
-  GRAPQL_JSON_PREFIX: 4,
-  COMMENT_PREFIX: 5,
+  GRAPHQL_JSON_PREFIX: 4,
+  COMMENT_PREFIX: 1,
   REQ_SEPARATOR: 9,
 };
 
@@ -124,6 +124,7 @@ module.exports = grammar({
     status_code: (_) => /[1-5]\d{2}/,
     status_text: (_) =>
       /(Continue|Switching Protocols|Processing|OK|Created|Accepted|Non-Authoritative Information|No Content|Reset Content|Partial Content|Multi-Status|Already Reported|IM Used|Multiple Choices|Moved Permanently|Found|See Other|Not Modified|Use Proxy|Switch Proxy|Temporary Redirect|Permanent Redirect|Bad Request|Unauthorized|Payment Required|Forbidden|Not Found|Method Not Allowed|Not Acceptable|Proxy Authentication Required|Request Timeout|Conflict|Gone|Length Required|Precondition Failed|Payload Too Large|URI Too Long|Unsupported Media Type|Range Not Satisfiable|Expectation Failed|I'm a teapot|Misdirected Request|Unprocessable Entity|Locked|Failed Dependency|Too Early|Upgrade Required|Precondition Required|Too Many Requests|Request Header Fields Too Large|Unavailable For Legal Reasons|Internal Server Error|Not Implemented|Bad Gateway|Service Unavailable|Gateway Timeout|HTTP Version Not Supported|Variant Also Negotiates|Insufficient Storage|Loop Detected|Not Extended|Network Authentication Required)/,
+
     __body: ($) =>
       seq(
         repeat1($._blank_line),
@@ -149,6 +150,7 @@ module.exports = grammar({
           ),
         ),
       ),
+
     response: ($) =>
       prec.right(
         seq(
@@ -281,29 +283,36 @@ module.exports = grammar({
 
     graphql_body: ($) =>
       prec.right(
-        seq($.graphql_data, optional(alias($.graphql_json_body, $.json_body))),
+        seq(
+          $.graphql_data,
+          optional(alias($.graphql_json_body, $.json_body)),
+          optional(alias($.graphql_external_body, $.external_body)),
+        ),
       ),
     graphql_data: ($) =>
       seq(
         token(
           prec(
             PREC.BODY_PREFIX,
-            seq(choice("query", "mutation"), WS, /.*\{/, NL),
+            seq(choice("query", "mutation"), WS, /.*[\{\(]\s*/, NL),
           ),
         ),
         $._raw_body,
       ),
     graphql_json_body: ($) =>
-      seq(token(prec(PREC.GRAPQL_JSON_PREFIX, /[{\[]\s+/)), $._raw_body),
+      seq(token(prec(PREC.GRAPHQL_JSON_PREFIX, /[{\[]\s*/)), $._raw_body),
+
+    graphql_external_body: ($) =>
+      seq(
+        token(prec(PREC.GRAPHQL_JSON_PREFIX, "<")),
+        WS,
+        field("path", $.path),
+        NL,
+      ),
 
     _external_body: ($) => seq($.external_body, NL),
     external_body: ($) =>
-      seq(
-        token(prec(PREC.BODY_PREFIX, "<")),
-        optional(seq("@", optional(field("name", $.identifier)))),
-        WS,
-        field("path", $.path),
-      ),
+      seq(token(prec(PREC.BODY_PREFIX, "<")), WS, field("path", $.path)),
 
     multipart_form_data: ($) =>
       prec.right(
@@ -312,7 +321,6 @@ module.exports = grammar({
           token(prec(1, LINE_TAIL)),
           repeat(
             choice(
-              $.comment,
               seq($.external_body, choice(WS, NL)),
               token(prec(2, /<[^\s@]/)),
               token(prec(2, "--")),
@@ -325,21 +333,9 @@ module.exports = grammar({
       ),
 
     raw_body: ($) =>
-      seq(
-        choice(
-          token(prec(1, seq(/.+/, NL))),
-          seq(COMMENT_PREFIX, $._not_comment),
-        ),
-        optional($._raw_body),
-      ),
+      seq(choice(token(prec(1, seq(/.+/, NL)))), optional($._raw_body)),
     _raw_body: ($) =>
-      seq(
-        choice(
-          token(prec(PREC.RAW_BODY, LINE_TAIL)),
-          seq(COMMENT_PREFIX, $._not_comment),
-        ),
-        optional($._raw_body),
-      ),
+      seq(choice(token(prec(PREC.RAW_BODY, LINE_TAIL))), optional($._raw_body)),
     _not_comment: (_) => token(seq(/[^@]*/, NL)),
 
     header_entity: (_) => /[\w\-]+/,