+++ NEWS.md
@@ -1,13 +1,22 @@
 # Kulala News
 
+## Version 5.3.3
+
+### Feature: support variables in redirect response path
+### Feature: do not display big responses + `max_response_size` config option
+### Feature: format json response on redirect
+
 ## Version 5.3.2
 
+### Enhancement: support url encoding in scheme, authority, path
+### Feature: support importing .graphql and .gql files
+### Enhancement: formatter wraps {{variables}} with quotes in json bodies
 ### Feature: add `# @secret` metadata to prompt for sensitive data
-### Enhancement: create `http-client.env.json` and `http-client.private.env.json` if not found
+### Feature: create `http-client.env.json` and `http-client.private.env.json` if not found
 ### Enhancement: add `Client Credentials` to all grant types
 ### Enhancement: expand variables in `Security.Auth` configs
 ### Enhancement: generate bug report on error and with `require("kulala").generate_bug_report()`
-### Enhancement: add `kulala_http` parser to markdown code blocks
+### Feature: support `kulala_http` parser in markdown code blocks
 ### Enhancement: update syntax highlighting for `kulala_http` parser
 ### Enhancement: add sorting options to [Formatter](getting-started/configuration-options.mdx)
 ### Enhancement: allow variables in curl and grpc flags